feat: SessionViewのDetail表示を別ファイルに移動し、レイアウト修正
